Add documentation on non-2xx status code handling
The documentation is not explicit about whether non-2xx status codes cause the request to fail and whether the user should handle 4xx codes like 429. However, from https://github.com/hashicorp/terraform-provider-http/blob/main/CHANGELOG.md#300-july-27-2022 it is clear that the request succeeds in this case and the user should in fact use a postcondition if they require a success response.